refactor: 디렉토리명 수정, 중복 타입 선언 제거
